+++ src/main/java/um/tesoreria/guarani/hexagonal/guarani/alumno/infrastructure/persistence/entity/AlumnoGuaraniEntity.java
@@ -4,6 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import um.tesoreria.guarani.hexagonal.guarani.persona.infrastructure.persistence.entity.PersonaGuaraniEntity;
 import um.tesoreria.guarani.hexagonal.guarani.propuesta.infrastructure.persistence.entity.PropuestaGuaraniEntity;
+import um.tesoreria.guarani.hexagonal.guarani.propuestaAspira.infrastructure.persistence.entity.PropuestaAspiraGuaraniEntity;
 import um.tesoreria.guarani.hexagonal.guarani.ubicacion.infrastructure.persistence.entity.UbicacionGuaraniEntity;
 
 import java.math.BigDecimal;
@@ -37,6 +38,13 @@ public class AlumnoGuaraniEntity {
     @OneToOne(optional = true)
     @JoinColumn(name = "ubicacion", referencedColumnName = "ubicacion", insertable = false, updatable = false)
     private UbicacionGuaraniEntity ubicacionRel;
+
+    @OneToOne(optional = true)
+    @JoinColumns({
+            @JoinColumn(name = "persona", referencedColumnName = "persona", insertable = false, updatable = false),
+            @JoinColumn(name = "propuesta", referencedColumnName = "propuesta", insertable = false, updatable = false)
+    })
+    private PropuestaAspiraGuaraniEntity propuestaAspiraRel;
 
     @Column(columnDefinition = "bpchar")
     private String modalidad;